Split variance resources into separate files
