+++ python-backend/app/api/mines/namespace/mines.py
@@ -15,9 +15,10 @@ from ..compliance.resources.compliance_article import ComplianceArticleResource
 from ..mine.resources.mine_basicinfo import MineBasicInfoResource
 from ..subscription.resources.subscription import MineSubscriptionResource, MineSubscriptionListResource
 from app.api.mines.mine.resources.mine_verified_status import MineVerifiedStatusResource
-from ..variances.resources.variance import (VarianceListResource, VarianceResource,
-                                            VarianceDocumentUploadResource,
-                                            VarianceUploadedDocumentsResource)
+from ..variances.resources.variance import VarianceResource
+from ..variances.resources.variance_list import VarianceListResource
+from ..variances.resources.variance_document_upload import VarianceDocumentUploadResource
+from ..variances.resources.variance_uploaded_documents import VarianceUploadedDocumentsResource
 from ..incidents.resources.mine_incidents import MineIncidentListResource, MineIncidentResource
 from ..incidents.resources.mine_incident_followup_types import MineIncidentFollowupTypeResource
